Fixed GetUserName(), Cleaned up source code
